+++ plugins/vmpayment/emspayideal/emspayideal.php
@@ -29,7 +29,7 @@ if (!class_exists('vmPSPlugin')) {
 }
 
 JLoader::registerNamespace('Emspay', JPATH_LIBRARIES . '/emspay');
-JImport('emspay.ems-php.vendor.autoload');
+JImport('emspay.ginger-php.vendor.autoload');
 JImport('emspay.emspayhelper');
 
 class plgVmPaymentEmspayideal extends EmspayVmPaymentPlugin
@@ -106,7 +106,7 @@ class plgVmPaymentEmspayideal extends EmspayVmPaymentPlugin
      */
     public function customInfoHTML()
     {
-        $issuers = $this->getGingerClient()->getIdealIssuers()->toArray();
+        $issuers = $this->getGingerClient()->getIdealIssuers();
         $html = '<select name="issuer" id="issuer" class="' . $this->_name . '">';
         foreach ($issuers as $issuer) {
             $html .= '<option value="' . $issuer['id'] . '">' . $issuer['name'] . "</option>";
@@ -203,34 +203,38 @@ class plgVmPaymentEmspayideal extends EmspayVmPaymentPlugin
         $webhook =$this->getWebhookUrl(intval($order['details']['BT']->virtuemart_paymentmethod_id));
         
         try {
-            $response = $this->getGingerClient()->createIdealOrder(
-                    $totalInCents,          // Amount in cents
-                    $currency_code_3,       // Currency
-                    $issuer,                // Issuer Id
-                    $description,           // Description
-                    $orderId,               // Merchant Order Id
-                    $returnUrl,             // Return URL
-                    null,                   // Expiration Period
-                    $customer->toArray(),   // Customer Information
-                    $plugin,                // Extra Information
-                    $webhook               // WebHook URL
-            );
+            $response = $this->getGingerClient()->createOrder([
+			'amount' => $totalInCents,                                   	// Amount in cents
+			'currency' => $currency_code_3,                              	// Currency
+			'transactions' => [
+				[
+					'payment_method' => 'ideal',                          // Payment method
+					'payment_method_details' => ['issuer_id' => $issuer]
+				]
+			],
+			'merchant_order_id' => $orderId,                             	// Merchant Order Id
+			'description' => $description,                               	// Description
+			'return_url' => $returnUrl,                                  	// Return URL
+			'customer' => $customer->toArray(),                          	// Customer Information
+			'extra' => ['plugin' => $plugin],                            	// Extra information
+			'webhook_url' => $webhook,                                   	// Webhook URL
+		]);
         } catch (\Exception $exception) {
             $html = "<p>" . JText::_("EMSPAY_LIB_ERROR_TRANSACTION") . "</p><p>Error: ".$exception->getMessage()."</p>";
             $this->processFalseOrderStatusResponse($html);
         }
 
-        if ($response->status()->isError()) {
-            $html = "<p>" . JText::_("EMSPAY_LIB_ERROR_TRANSACTION") . "</p><p>Error: ".$response->transactions()->current()->reason()->toString()."</p>";
+        if ($response['status'] == 'error') {
+            $html = "<p>" . JText::_("EMSPAY_LIB_ERROR_TRANSACTION") . "</p><p>Error: ".$response['transactions'][0]['reason']."</p>";
             $this->processFalseOrderStatusResponse($html);
         }
 
-        if (!$response->getId()) {
+        if (!$response['id']) {
             $html = "<p>" . JText::_("EMSPAY_LIB_ERROR_TRANSACTION") . "</p><p>Error: Response did not include id!</p>";
             $this->processFalseOrderStatusResponse($html);
         }
 
-        if (!$response->firstTransactionPaymentUrl()) {
+        if (!$response['transactions'][0]['payment_url']) {
             $html = "<p>" . JText::_("EMSPAY_LIB_ERROR_TRANSACTION") . "</p><p>Error: Response did not include payment url!</p>";
             $this->processFalseOrderStatusResponse($html);
         }
@@ -247,11 +251,11 @@ class plgVmPaymentEmspayideal extends EmspayVmPaymentPlugin
         $dbValues['email_currency'] = $email_currency;
         $dbValues['payment_order_total'] = $totalInPaymentCurrency['value'];
         $dbValues['tax_id'] = $method->tax_id;
-        $dbValues['ginger_order_id'] = $response->id()->toString();
+        $dbValues['ginger_order_id'] = $response['id'];
 
         $this->storePSPluginInternalData($dbValues);
 
-        JFactory::getApplication()->redirect($response->firstTransactionPaymentUrl()->toString());
+        JFactory::getApplication()->redirect($response['transactions'][0]['payment_url']);
     }
 
     /**
@@ -280,12 +284,8 @@ class plgVmPaymentEmspayideal extends EmspayVmPaymentPlugin
 
         $gingerOrder = $this->getGingerClient()->getOrder(vRequest::get('order_id'));
 
-        if (!$gingerOrder instanceof GingerPayments\Payment\Order) {
-            return JFactory::getApplication()->enqueueMessage("Error: Some text!", 'error');
-        }
-
         $virtuemart_order_id = $this->getOrderIdByGingerOrder(vRequest::get('order_id'));
-        $statusSucceeded = $this->updateOrder($gingerOrder->getStatus(), $virtuemart_order_id);
+        $statusSucceeded = $this->updateOrder($gingerOrder['status'], $virtuemart_order_id);
  
         $html = "<p>" . EmspayHelper::getOrderDescription($virtuemart_order_id) . "</p>";
         
@@ -295,7 +295,7 @@ class plgVmPaymentEmspayideal extends EmspayVmPaymentPlugin
             $this->processFalseOrderStatusResponse($html);
         }
         
-        if ($gingerOrder->getStatus() === 'processing' && !$this->isProcessingOrderNotConfirmedRedirect()) {
+        if ($gingerOrder['status'] === 'processing' && !$this->isProcessingOrderNotConfirmedRedirect()) {
             $box = '
                 jQuery(document).ready(function($) {
                     var fallback_url = \''. JURI::root() .'?option=com_virtuemart&view=pluginresponse&task=pluginresponsereceived&pm='.vRequest::getInt('pm').'&project_id='.vRequest::getInt('project_id').'&order_id='.vRequest::get('order_id').'&no_confirmation_redirect=1\';
@@ -333,7 +333,7 @@ class plgVmPaymentEmspayideal extends EmspayVmPaymentPlugin
         }
       
         if ($statusSucceeded === false) {
-            switch ($gingerOrder->getStatus()) {
+            switch ($gingerOrder['status']) {
                 case 'expired':
                     $html .=  "<p>" . JText::_("EMSPAY_LIB_ERROR_STATUS_EXPIRED") . "</p>";
                     break;
@@ -381,16 +381,8 @@ class plgVmPaymentEmspayideal extends EmspayVmPaymentPlugin
             return false;
         }
         
-        $gingerOrder = $this->getGingerClient()->getOrder(vRequest::get('order_id'));
         $response = ['status' => 'not valid order','redirect' => true];
         
-        if ($gingerOrder instanceof GingerPayments\Payment\Order) {
-            if ($gingerOrder->getStatus() == 'processing') {
-                $response = ['status' => $gingerOrder->getStatus(),'redirect' => false];
-            } else {
-                $response = ['status' => $gingerOrder->getStatus(),'redirect' => true];
-            }
-        }
         die(json_encode($response));
     }
     
@@ -417,13 +409,9 @@ class plgVmPaymentEmspayideal extends EmspayVmPaymentPlugin
 
         $gingerOrder = $this->getGingerClient()->getOrder($input['order_id']);
 
-        if (!$gingerOrder instanceof GingerPayments\Payment\Order) {
-            exit("Invalid order");
-        }
-
         $virtuemart_order_id = $this->getOrderIdByGingerOrder($input['order_id']);
 
-        $this->updateOrder($gingerOrder->getStatus(), $virtuemart_order_id);
+        $this->updateOrder($gingerOrder['status'], $virtuemart_order_id);
 
         exit();
     }